requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
boto3==1.38.0
python-dotenv==1.1.0
